Move system support specs to separate directory
