+++ spec/support/system/allows_interacting_with_tickets_examples.rb
